Tests/LibWeb/Text/expected/hit_testing/block-element-inside-inline-element.txt

<#text>
index: 2
<BODY>
---
<#text>
index: 1
<B id="b1">
---
<#text>
index: 2
<DIV id="d1">
---
<#text>
index: 3
<B id="b1">
---
